+++ CHANGELOG.md
@@ -5,8 +5,14 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Changed
+- BREAKING: Introduced proper (Result) error handling for `transform_all_roles_to_config` ([#282]).
+- BREAKING: `Configuration::compute_*` are now invoked even when `config` field is not provided on `Role`/`RoleGroup` ([#282]).
+  - `CommonConfiguration::config` is no longer `Option`al
+  - `Role::config` is no longer `Option`al
+  - `RoleGroup::config` is no longer `Option`al
 - Fixed `cli::Command` including developer-facing docs in `--help` output ([#283])
 
+[#282]: https://github.com/stackabletech/operator-rs/pull/282
 [#283]: https://github.com/stackabletech/operator-rs/pull/283
 
 ## [0.6.0] - 2021-12-13